Dev -> main

+++ Pico/Home/HomeUserCardViewController.swift
@@ -251,8 +251,8 @@ final class HomeUserCardViewController: UIViewController {
     }
     
     private func configButtons() {
-        likeButton.addTarget(self, action: #selector(tappedLikeButton), for: .touchUpInside)
-        disLikeButton.addTarget(self, action: #selector(tappedDisLikeButton), for: .touchUpInside)
+        likeButton.addTarget(self, action: #selector(tappedLikeButtonObjc), for: .touchUpInside)
+        disLikeButton.addTarget(self, action: #selector(tappedDisLikeButtonObjc), for: .touchUpInside)
         pickBackButton.addTarget(self, action: #selector(tappedPickBackButton), for: .touchUpInside)
     }
     
@@ -275,7 +275,7 @@ final class HomeUserCardViewController: UIViewController {
         panGesture = UIPanGestureRecognizer(target: self, action: #selector(touchGesture(_:)))
         self.view.addGestureRecognizer(panGesture)
     }
-        
+    
     private func loadUserImages() {
         for (index, url) in user.imageURLs.enumerated() {
             let imageView = UIImageView()
@@ -337,15 +337,24 @@ final class HomeUserCardViewController: UIViewController {
                     HomeUserCardViewModel.cardCounting = -1
                     homeViewController?.addUserCards()
                 }
-
-                NotificationService.shared.sendNotification(userId: user.id, sendUserName: currentUser.nickName, notiType: .like)
-                guard let myMbti = MBTIType(rawValue: currentUser.mbti) else { return }
-                let noti = Noti(receiveId: user.id, sendId: currentUser.userId, name: currentUser.nickName, birth: currentUser.birth, imageUrl: currentUser.imageURL, notiType: .like, mbti: myMbti, createDate: Date().timeIntervalSince1970)
-                FirestoreService.shared.saveDocument(collectionId: .notifications, data: noti)
-
+                homeViewController?.removedView.append(view)
+                
+                if currentUser.userId.prefix(4) != Bundle.main.testId {
+                    viewModel.checkYouLikeMe(user.id, currentUser.userId) { [self] result in
+                        if result {
+                            viewModel.saveLikeData(receiveUserInfo: user, likeType: .matching)
+                            viewModel.updateMatcingData(user.id)
+                            viewModel.notificationServiceForPartner(.matching, .matching, user: user, currentUser: currentUser)
+                            viewModel.notificationServiceForMe(.matching, .matching, user: user, currentUser: currentUser)
+                            homeViewController?.removedView.removeLast()
+                        } else {
+                            viewModel.saveLikeData(receiveUserInfo: user, likeType: .like)
+                            viewModel.notificationServiceForPartner(.like, .like, user: user, currentUser: currentUser)
+                        }
+                    }
+                }
+                
                 UIView.animate(withDuration: 0.5) { [self] in
-                    viewModel.saveLikeData(receiveUserInfo: user, likeType: .like)
-                    homeViewController?.removedView.append(view)
                     view.center.x += 1000
                     homeViewController?.likeLabel.alpha = 0
                 }
@@ -355,7 +364,9 @@ final class HomeUserCardViewController: UIViewController {
                     HomeUserCardViewModel.cardCounting = -1
                     homeViewController?.addUserCards()
                 }
-                self.viewModel.saveLikeData(receiveUserInfo: user, likeType: .dislike)
+                if currentUser.userId.prefix(4) != Bundle.main.testId {
+                    self.viewModel.saveLikeData(receiveUserInfo: user, likeType: .dislike)
+                }
                 UIView.animate(withDuration: 0.5) { [self] in
                     homeViewController?.removedView.append(view)
                     view.center.x -= 1000
@@ -373,22 +384,34 @@ final class HomeUserCardViewController: UIViewController {
             break
         }
     }
-    @objc func tappedLikeButton() {
+    @objc func tappedLikeButtonObjc() {
+        tappedLikeButton()
+    }
+    
+    func tappedLikeButton() {
         self.homeViewController?.removedView.append(self.view)
         self.homeViewController?.likeLabel.alpha = 1
-        self.viewModel.saveLikeData(receiveUserInfo: user, likeType: .like)
-        
         HomeUserCardViewModel.cardCounting += 1
         if HomeUserCardViewModel.cardCounting == 3 {
             HomeUserCardViewModel.cardCounting = -1
             homeViewController?.addUserCards()
         }
+        if currentUser.userId.prefix(4) != Bundle.main.testId {
+            viewModel.checkYouLikeMe(user.id, currentUser.userId) { [self] result in
+                if result {
+                    viewModel.saveLikeData(receiveUserInfo: user, likeType: .matching)
+                    viewModel.updateMatcingData(user.id)
+                    viewModel.notificationServiceForPartner(.matching, .matching, user: user, currentUser: currentUser)
+                    viewModel.notificationServiceForMe(.matching, .matching, user: user, currentUser: currentUser)
+                    homeViewController?.removedView.removeLast()
 
-        NotificationService.shared.sendNotification(userId: user.id, sendUserName: currentUser.nickName, notiType: .like)
-        guard let myMbti = MBTIType(rawValue: currentUser.mbti) else { return }
-        let noti = Noti(receiveId: user.id, sendId: currentUser.userId, name: currentUser.nickName, birth: currentUser.birth, imageUrl: currentUser.imageURL, notiType: .like, mbti: myMbti, createDate: Date().timeIntervalSince1970)
-        FirestoreService.shared.saveDocument(collectionId: .notifications, data: noti)
-
+                } else {
+                    viewModel.saveLikeData(receiveUserInfo: user, likeType: .like)
+                    viewModel.notificationServiceForPartner(.like, .like, user: user, currentUser: currentUser)
+                }
+            }
+        }
+        
         UIView.animate(withDuration: 0.5) {
             self.view.center.x += 1000
         } completion: { _ in
@@ -396,15 +419,23 @@ final class HomeUserCardViewController: UIViewController {
         }
     }
     
-    @objc func tappedDisLikeButton() {
+    @objc func tappedDisLikeButtonObjc() {
+        tappedDisLikeButton()
+    }
+    
+    func tappedDisLikeButton() {
         self.homeViewController?.removedView.append(self.view)
         self.homeViewController?.passLabel.alpha = 1
-        self.viewModel.saveLikeData(receiveUserInfo: user, likeType: .dislike)
         HomeUserCardViewModel.cardCounting += 1
         if HomeUserCardViewModel.cardCounting == 3 {
             HomeUserCardViewModel.cardCounting = -1
             homeViewController?.addUserCards()
         }
+        
+        if currentUser.userId.prefix(4) != Bundle.main.testId {
+            self.viewModel.saveLikeData(receiveUserInfo: user, likeType: .dislike)
+        }
+        
         UIView.animate(withDuration: 0.5) {
             self.view.center.x -= 1000
         } completion: { _ in
@@ -425,7 +456,9 @@ final class HomeUserCardViewController: UIViewController {
                     let remainingChu = UserDefaultsManager.shared.getChuCount()
                     if remainingChu >= 10 {
                         viewModel.purchaseChu(currentChu: remainingChu, purchaseChu: 10)
-                        self.viewModel.deleteLikeData()
+                        if currentUser.userId.prefix(4) != Bundle.main.testId {
+                            self.viewModel.deleteLikeData()
+                        }
                         UIView.animate(withDuration: 0.3) {
                             lastView.center = self.view.center
                             lastView.transform = CGAffineTransform(rotationAngle: 0)
@@ -440,7 +473,7 @@ final class HomeUserCardViewController: UIViewController {
                 }
             )
         } else {
-            showCustomAlert(alertType: .onlyConfirm, titleText: "첫번째 추천입니다.", messageText: "", confirmButtonText: "확인")
+            showCustomAlert(alertType: .onlyConfirm, titleText: "이전 친구를 찾을 수 없습니다.", messageText: "💡 매칭된 상대는 되돌릴 수 없습니다.", confirmButtonText: "확인")
         }
     }
     
@@ -463,6 +496,12 @@ final class HomeUserCardViewController: UIViewController {
     @objc func tappedInfoButton() {
         let viewController = UserDetailViewController()
         viewController.viewModel = UserDetailViewModel(user: user, isHome: true)
+        viewController.onLike = { [weak self] in
+            self?.tappedLikeButtonObjc()
+        }
+        viewController.onDisLike = { [weak self] in
+            self?.tappedDisLikeButton()
+        }
         navigationController?.pushViewController(viewController, animated: true)
     }
